requirements.txt
alabaster==0.7.10
Babel==2.5.3
backcall==0.1.0
bleach==2.1.3
boto==2.40.0
certifi==2018.4.16
chardet==3.0.4
classify==0.5.0
colorama==0.3.9
cycler==0.10.0
decorator==4.2.1
defusedxml==0.5.0
dj-database-url==0.4.1
Django==2.0
django-allauth==0.38.0
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-js-asset==1.1.0
django-lineage==0.2.0
django-markdown==0.8.4
django-mysql==2.4.1
django-simple-pagination==1.1.4
django-storages==1.4.1
django-webpack-loader==0.6.0
djangorestframework==3.8.2
docutils==0.14
entrypoints==0.2.3
gunicorn==19.6.0
html5lib==1.0.1
idna==2.6
imagesize==1.0.0
ipykernel==4.8.2
ipyparallel==6.1.1
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lineage==1.0.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.2.3
microurl==0.1.1
mistune==0.8.3
mysqlclient==1.3.12
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==5.4.1
npm==0.1.1
numpy==1.15.1
oauthlib==2.1.0
olefile==0.46
optional-django==0.1.0
packaging==17.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.1.1
patsy==0.5.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
psycopg2==2.7.1
Pygments==2.2.0
PyMySQL==0.8.0
pyparsing==2.2.0
python-dateutil==2.7.2
python3-openid==3.1.0
pytz==2018.3
pywinpty==0.5.1
PyYAML==3.13
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.10.0
requests-oauthlib==1.0.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinxcontrib-websupport==1.0.1
statsmodels==0.9.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.1
traitlets==4.3.2
transliterate==1.10.2
urllib3==1.22
virtualenv==16.1.0
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.1
widgetsnbextension==3.2.0
xlrd==1.1.0
